requirements.txt
albumentations
configargparse
matplotlib
monai
nibabel
numpy
opencv-python
pandas
pathlib
polyaxon-client
lightning==2.0.0
scikit-learn
SimpleITK
torchgeometry
#torchvision==0.15.2
#torchaudio==2.0.2
torchmetrics==1.0.1
#torchtext==0.15.2
tqdm
wandb
seaborn